% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lossFunction.r
\name{calc_deviance}
\alias{calc_deviance}
\title{calc_deviance}
\usage{
calc_deviance(actual, predicted, family = "poisson", ...)
}
\arguments{
\item{actual}{actual data values}

\item{predicted}{predicted data values}

\item{family}{default value is poisson}

\item{\dots}{can receive parameters accepted by dismo::calc.deviance}
}
\value{
error between actual versus prediction
}
\description{
Functions for calculating deviance
}
\details{
loss functions written for use in cv.glmmLasso
}
\author{
Pirapong Jitngamplang, Jared Lander
}
