% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lossFunction.r
\name{calc_logloss}
\alias{calc_logloss}
\title{calc_logloss}
\usage{
calc_logloss(actual, predicted)
}
\arguments{
\item{actual}{actual data values}

\item{predicted}{predicted data values}
}
\value{
error between actual versus prediction
}
\description{
Functions for calculating logloss
}
\details{
Loss functions written for use in cv.glmmLasso
}
\author{
Pirapong Jitngamplang, Jared Lander
}
