% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lossFunction.r
\name{calc_mse}
\alias{calc_mse}
\title{calc_mse}
\usage{
calc_mse(actual, predicted)
}
\arguments{
\item{actual}{actual data values}

\item{predicted}{predicted data values}
}
\value{
error between actual versus prediction
}
\description{
Function for calculating mse
}
\details{
Loss functions written for use in cv.glmmLasso
}
\author{
Pirapong Jitngamplang, Jared Lander
}
