% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lossFunction.r
\name{calc_multilogloss}
\alias{calc_multilogloss}
\title{calc_multilogloss}
\usage{
calc_multilogloss(actual, predicted)
}
\arguments{
\item{actual}{actual data values}

\item{predicted}{predicted data values}
}
\value{
error between actual versus prediction
}
\description{
Function for calculating multilogloss
}
\details{
loss functions written for use in cv.glmmLasso
}
\author{
Pirapong Jitngamplang, Jared Lander
}
