% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLambda.R
\name{computeLambdaMax}
\alias{computeLambdaMax}
\title{computeLambdaMax}
\usage{
computeLambdaMax(fix, rnd, data, scale = TRUE)
}
\arguments{
\item{fix}{A two-sided linear formula object describing the fixed-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right. For categorical covariables use as.factor(.) in the formula. Note, that the corresponding dummies are treated as a group and are updated blockwise}

\item{rnd}{A two-sided linear formula object describing the random-effects part of the model, with the grouping factor on the left of a ~ operator and the random terms, separated by + operators, on the right; aternatively, the random effects design matrix can be given directly (with suitable column names). If set to NULL, no random effects are included.}

\item{data}{The data frame containing the variables named in formula.}

\item{scale}{default value is true}
}
\value{
returns the lambdaMax value based on given dataset
}
\description{
compute the maximum lambda value based on given dataset
}
\details{
lambdaMax is computed based on  the coordinate descent algorithm from this paper: Friedman, Jerome, Trevor Hastie, and Rob Tibshirani. "Regularization paths for generalized linear models via coordinate descent."
}
\author{
Pirapong Jitngamplang, Jared Lander
}
