% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmmLasso.R
\name{cv.glmmLasso}
\alias{cv.glmmLasso}
\title{cv.glmmLasso}
\usage{
cv.glmmLasso(fix, rnd, data, family = gaussian(link = "identity"),
  kfold = 5, lambdas = NULL, nlambdas = 100,
  lambda.min.ratio = ifelse(nobs < nvars, 0.01, 1e-04), loss,
  lambda.final = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{fix}{A two-sided linear formula object describing the fixed-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right. For categorical covariables use as.factor(.) in the formula. Note, that the corresponding dummies are treated as a group and are updated blockwise}

\item{rnd}{A two-sided linear formula object describing the random-effects part of the model, with the grouping factor on the left of a ~ operator and the random terms, separated by + operators, on the right; aternatively, the random effects design matrix can be given directly (with suitable column names). If set to NULL, no random effects are included.}

\item{data}{The data frame containing the variables named in formula.}

\item{family}{A GLM family, see \code{\link[=glm]{glm()}} and \code{\link[=family]{family()}}. Also ordinal response models can be fitted: use family=acat() and family=cumulative() for the fitting of an adjacent category or cumulative model, respectively. If family is missing then a linear mixed model is fit; otherwise a generalized linear mixed model is fit.}

\item{lambdas}{Optional user-supplied lambda sequence; default is NULL, and glmmLasso_MultLambdas chooses its own sequence}

\item{lambda.min.ratio}{Smallest value for lambda, as a fraction of lambda.max, the (data derived) entry value (i.e. the smallest value for which all coefficients are zero). The default depends on the sample size nobs relative to the number of variables nvars. If nobs > nvars, the default is 0.0001, close to zero. If nobs < nvars, the default is 0.01.}

\item{loss}{Loss function used to calculate error, default values is based on family - 'gaussian' = calc_mse, 'binomial' = calc_logloss, 'multinomial' = calc_multilogloss, 'poisson' = calc_deviance}

\item{lambda.final}{Choice for final model to use lambda.1se or lambda.min, default is lambda.1se}

\item{nlambda}{The number of lambdas values, default value is 100 if lambdas is not user-supplied}
}
\value{
A list of cross-validation values including lambdas, cvm, cvsd, cvup, cvlo, glmmLasso.final, lambda.min, lambda.1se
}
\description{
Does k-fold cross validation for glmmLasso
}
\details{
Build multiple models given a sequence of lambda values
}
\examples{
data("soccer", package = "glmmLasso")
soccer[,c(4,5,9:16)]<-scale(soccer[,c(4,5,9:16)],center=TRUE,scale=TRUE)
soccer<-data.frame(soccer)
fix = points ~ transfer.spendings + ave.unfair.score + ball.possession + tackles + ave.attend + sold.out
rnd = list(team=~1)
cv.glmmLasso(fix = points ~ transfer.spendings + ave.unfair.score + ball.possession + tackles + ave.attend + sold.out, rnd = list(team=~1), data = soccer, family= gaussian(link = "identity"), kfold = 5, lambda.final= 'lambda.1se')
}
\author{
Pirapong Jitngamplang, Jared Lander
}
